Module Octez_protocol_compiler_compat.Compiler_libs

val default_persistent_signature_loader : allow_hidden:bool -> unit_name:string -> Persistent_env.Persistent_signature.t option

val override_persistent_signature_loader : (allow_hidden:bool -> unit_name:string -> Persistent_env.Persistent_signature.t option) -> unit

set_persistent_signature_loader f calls `Persistent_env.Persistent_signature.load := f`, but handles the necessary wrapping for it to work on different version of OCaml.

val make_persistent_signature : filename:string -> cmi:Cmi_format.cmi_infos -> Persistent_env.Persistent_signature.t

make_persistent_signature ~filename ~cmi wraps `Persistent_signature.t` to be compatible with multiple version of OCaml.

val compunit_name : Cmo_format.compilation_unit -> string

compunit_name cu returns the name of the compilation unit

val mark_attribute_used : Parsetree.attribute -> unit

mark_attribute_used attr marks the attribute used, so that it doesn't trigger warning 53. See Packer.check_syntax.